Top 3 best categories based on the confidence.
The list that the Google API delivers is ordered from highest to lowest confidence, so you only need to limit it to obtaine the first 3 values ​​at most. Finally, you are getting the most specific category for each of your results, and if the category turns out to be "Other", you need to simply get the more general you category on the left of "Other".